Make shake to help as new root for the whole app.
Refactored shake to help to its own view controller and made the new view controller the new root for the whole app, because the shake to help was handled by the splash view controller which should not be responsible for that as well.
