+++ ostelco-ios-client/Storyboards/Splash.storyboard
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14490.70" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="MAv-rn-oJO">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14490.70" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="wmV-iq-aAT">
     <device id="retina5_9" orientation="portrait">
         <adaptation id="fullscreen"/>
     </device>
@@ -37,6 +37,35 @@
                 <placeholder placeholderIdentifier="IBFirstResponder" id="GEL-B7-INL" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
             <point key="canvasLocation" x="-1762" y="-336"/>
+        </scene>
+        <!--Shake To Help View Controller-->
+        <scene sceneID="4M5-Hf-aR9">
+            <objects>
+                <viewController id="wmV-iq-aAT" customClass="ShakeToHelpViewController" customModule="Oya_Development_app" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="84d-Uy-9VX">
+                        <rect key="frame" x="0.0" y="0.0" width="375" height="812"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <containerView opaque="NO" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="6dH-hN-J46">
+                                <rect key="frame" x="0.0" y="0.0" width="375" height="812"/>
+                                <connections>
+                                    <segue destination="MAv-rn-oJO" kind="embed" id="lZk-cS-06E"/>
+                                </connections>
+                            </containerView>
+                        </subviews>
+                        <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                        <constraints>
+                            <constraint firstAttribute="bottom" secondItem="6dH-hN-J46" secondAttribute="bottom" id="0Uz-AO-rcW"/>
+                            <constraint firstAttribute="trailing" secondItem="6dH-hN-J46" secondAttribute="trailing" id="9oO-h4-4pk"/>
+                            <constraint firstItem="6dH-hN-J46" firstAttribute="top" secondItem="84d-Uy-9VX" secondAttribute="top" id="EnQ-gZ-QgK"/>
+                            <constraint firstItem="6dH-hN-J46" firstAttribute="leading" secondItem="84d-Uy-9VX" secondAttribute="leading" id="k73-1D-HFD"/>
+                        </constraints>
+                        <viewLayoutGuide key="safeArea" id="BVB-Oj-uk6"/>
+                    </view>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="yxl-Gd-QDy" userLabel="First Responder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="-1762" y="-1200"/>
         </scene>
         <!--Address-->
         <scene sceneID="Jvt-4A-kf4">
